+++ libs/sabi-core/src/auth/v4/key/hmac_factory.rs
@@ -1,6 +1,5 @@
-use crate::auth::v4::key::CanGenerateHmac;
+use crate::auth::v4::key::{CanGenerateHmac, OutputHmac};
 use crate::verbs::AsBytes;
-use hex::ToHex;
 use hmac::{Hmac, Mac};
 use sha2::Sha256;
 
@@ -18,22 +17,3 @@ pub trait HmacFactory: CanGenerateHmac {
 }
 
 impl<A> HmacFactory for A where A: CanGenerateHmac {}
-
-pub struct OutputHmac {
-    code: Vec<u8>,
-}
-
-impl OutputHmac {
-    pub fn code(self) -> Vec<u8> {
-        self.code
-    }
-    pub fn to_hex(&self) -> String {
-        self.code.encode_hex()
-    }
-}
-
-impl AsBytes for OutputHmac {
-    fn as_bytes(&self) -> &[u8] {
-        &self.code
-    }
-}
