+++ libs/sabi-core/src/auth/v4/key/signing_key.rs
@@ -16,7 +16,7 @@ impl SigningKey {
             .hmac(&scope.service)
             .hmac(&scope.termination);
 
-        SigningKey(hmac.code())
+        SigningKey(hmac.code)
     }
 
     pub fn to_hex(&self) -> String {
